decode pms5003 data

+++ esp-device/main/include/dto.h
@@ -6,9 +6,9 @@
 typedef struct {
     uint32_t device_timestamp;
     char device_name[20];
-    uint8_t pm10_standard;
-    uint8_t pm25_standard;
-    uint8_t pm100_standard;
+    uint16_t pm10_standard;
+    uint16_t pm25_standard;
+    uint16_t pm100_standard;
 } pms5003_measurement;
 
 #endif
